move instance affinity group list to settings tab
