+++ app/table/cells/DescriptionCell.tsx
@@ -9,5 +9,7 @@
 import { EmptyCell } from '~/table/cells/EmptyCell'
 import { Truncate } from '~/ui/lib/Truncate'
 
-export const DescriptionCell = ({ text }: { text?: string }) =>
-  text ? <Truncate text={text} maxLength={48} /> : <EmptyCell />
+export type Props = { text?: string; maxLength?: number }
+
+export const DescriptionCell = ({ text, maxLength = 48 }: Props) =>
+  text ? <Truncate text={text} maxLength={maxLength} /> : <EmptyCell />
